+++ src/core/contextType.ts
@@ -5234,6 +5234,12 @@ export interface KAPLAYCtx {
         sides: number,
         startAngle: number,
     ): Vec2[];
+    createCogPolygon(
+        radius1: number,
+        radius2: number,
+        sides: number,
+        startAngle: number,
+    ): Vec2[];
     /**
      * Check if a line and a point intersect.
      *